requirements.txt
blinker==1.8.2
cachelib==0.13.0
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.0.8
Werkzeug==3.0.3
